1.5.7
DLL interface optimization and small improvements.

+++ BDMatchCore/headers/multithreading.h
@@ -96,7 +96,7 @@ private:
 		std::mutex mtx_;
 		std::condition_variable cond_;
 		std::condition_variable fin_;
-		std::atomic<int> task_num_ = 0;
+		std::atomic<size_t> task_num_ = 0;
 		std::atomic_flag *keep_processing = nullptr;
 		bool is_shutdown_ = false;
 		std::queue<std::function<void()>> tasks_;
